# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Expresso café — API",
    "version": "1.0.0",
    "description": "Dinheiro em centavos. Sessão por cookie httpOnly."
  },
  "components": {
    "schemas": {
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "administrador": {
            "type": "boolean",
            "description": "Administrador do sistema: só usa o painel de empresas"
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "administrador"
        ],
        "additionalProperties": false
      },
...
      "Entrar": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "senha"
        ],
        "additionalProperties": false
      },
...
      "MudarSenha": {
        "type": "object",
        "properties": {
          "senhaAtual": {
            "type": "string"
          },
          "novaSenha": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "senhaAtual",
          "novaSenha"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "expresso_sessao"
      }
    }
  },
  "security": [
    {
      "cookie": []
    }
  ],
  "paths": {
    "/sessao": {
      "post": {
        "summary": "Entrar (grava o cookie de sessão)",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Entrar"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Quem está logado",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Sair",
        "security": [],
        "responses": {
          "204": {
            "description": "Sem conteúdo"
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/sessao/senha": {
      "put": {
        "summary": "Trocar a própria senha",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MudarSenha"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Sem conteúdo"
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/configuracao": {
      "get": {
        "summary": "Configurações do estabelecimento (sem login, os padrões)",
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Configuracao"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Salvar configurações",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalvarConfiguracao"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Configuracao"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/usuarios": {
      "get": {
        "summary": "Pessoas com acesso",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Dar acesso a alguém",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovoUsuario"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/dias": {
      "get": {
        "summary": "Dias de venda",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumoDia"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Abrir dia de venda",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbrirDia"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaVenda"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/dias/sugestao-producao": {
      "get": {
        "summary": "Último, média e sobra média por produto",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SugestaoProducao"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/dias/relatorio": {
      "get": {
        "summary": "Fotos dos últimos dias fechados",
        "parameters": [
          {
            "name": "dias",
            "in": "query",
            "required": false,
            "description": "Quantos dias (1 a 26, padrão 8)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiaFechado"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/dias/{id}/painel": {
      "get": {
        "summary": "Início do dia",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PainelDoDia"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/dias/{id}/fechamento": {
      "post": {
        "summary": "Fechar o dia (somente leitura depois)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sem conteúdo"
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/dias/{id}/pedidos": {
      "get": {
        "summary": "Pedidos do dia",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "retirada",
            "in": "query",
            "required": false,
            "description": "reservado | retirado | cancelado",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "busca",
            "in": "query",
            "required": false,
            "description": "Nome, telefone ou número",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pedido"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Confirmar reserva (409 com o máximo se não couber)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovoPedido"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pedido"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/dias/{id}/producao": {
      "get": {
        "summary": "Produção, reservados e vendidos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EstoqueDoProduto"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Salvar produção (409 abaixo do reservado)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalvarProducao"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Sem conteúdo"
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/dias/{id}/producao/historico": {
      "get": {
        "summary": "Quem mudou a produção",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlteracaoProducao"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/dias/{id}/espera": {
      "get": {
        "summary": "Lista de espera do dia",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntradaEspera"
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Entrar na lista de espera",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NovaEspera"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntradaEspera"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/dias/{id}/eventos": {
      "get": {
        "summary": "Tempo real (text/event-stream): disponibilidade-mudou, unidades-liberadas",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sem conteúdo"
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/pedidos/{id}": {
      "get": {
        "summary": "Detalhe do pedido",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pedido"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Editar itens (409 se não couber)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditarPedido"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pedido"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/pedidos/{id}/retirada": {
      "post": {
        "summary": "Marcar como retirado",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pedido"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/pedidos/{id}/pagamento": {
      "put": {
        "summary": "Registrar pagamento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pagamento"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pedido"
                }
              }
            }
          },
          "4XX": {
            "description": "Erro com a mensagem para o balcão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
  }
}